# src/main/resources/users/api.yaml
openapi: 3.0.0
info:
  version: 1.0.2
  title: User Data Manager
  description: 'Just an example of CRUD API'
  contact:
    name: SkyWaet
    url: https://github.com/SkyWaet
    email: SkyWaet@gmail.com
components:
  schemas:
    UsersListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserShortDTO'
        pageable:
          $ref: '../pageable/pageable.yml#/components/schemas/Pageable'
        total_pages:
          type: integer
        total_elements:
          type: integer
        last:
          type: boolean
        first:
          type: boolean
        size:
          type: integer
        number_of_elements:
          type: integer
        number:
          type: integer
        sort:
          $ref: '../pageable/pageable.yml#/components/schemas/Sort'
        empty:
          type: boolean

    UserShortDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
        login:
          type: string
          minLength: 6
          maxLength: 20
          description: login of the user
        password:
          type: string
          minLength: 8
          maxLength: 20
          format: password
          description: password of the user
        name:
          type: string
          maxLength: 50
          description: user name
        surname:
          type: string
          maxLength: 100
          description: user surname
        patronymic:
          type: string
          maxLength: 100
          description: user patronymic
        email:
          type: string
          format: email
          nullable: true
          description: user email
      required:
        - login
        - password

    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
        login:
          type: string
          minLength: 6
          maxLength: 20
          description: login of the user
        password:
          type: string
          minLength: 8
          maxLength: 20
          format: password
          description: password of the user
        name:
          type: string
          maxLength: 50
          description: user name
        surname:
          type: string
          maxLength: 100
          description: user surname
        patronymic:
          type: string
          maxLength: 100
          description: user patronymic
        email:
          type: string
          format: email
          nullable: true
          description: user email
        desc:
          type: string
          nullable: true
          maxLength: 1536
          description: description of user
        vacations:
          type: array
          items:
            $ref: '#/components/schemas/VacationDTO'
          required: true
      required:
        - login
        - password

    VacationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time


  examples:
    JohnReq:
      value:
        id: null
        login: john
        password: john123
        name: John
        surname: Johnson
        patronymic: James
        email: john@gmail.com
        description: null
    MaryReq:
      value:
        id: null
        login: mary
        password: mary123
        name: Mary
        surname: Smith
        patronymic: James
        email: null
        description: "Hello, I am Mary. I love coding!"
    JohnRes:
      value:
        id: 0
        login: john
        password: john123
        name: John
        surname: Johnson
        patronymic: James
        email: john@gmail.com
        description: null
    MaryRes:
      value:
        id: 1
        login: mary
        password: mary123
        name: Mary
        surname: Smith
        patronymic: James
        email: null
        description: "Hello, I am Mary. I love coding!"

  requestBodies:
    userBody:
      required: true
      description: A JSON object containing new user data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserDTO'
          examples:
            john:
              $ref: '#/components/examples/JohnReq'
            mary:
              $ref: '#/components/examples/MaryReq'

paths:
  /users:
    get:
      summary: Gets list of users
      parameters:
        - in: query
          name: pageIndex
          schema:
            type: integer
            minimum: 0
            default: 0
          description: number of desired page
        - in: query
          name: numElements
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            description: amount of elements to return


      responses:
        '200':
          description: Gets an array of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
              example:
                - id: 0
                  login: John
                  password: john123
                  email: john@gmail.com
                - id: 1
                  login: Mary
                  password: mary456
                  email: null
    post:
      summary: Creates new user
      requestBody:
        $ref: '#/components/requestBodies/userBody'

      responses:
        '200':
          description: Returns a copy of created user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
              examples:
                john:
                  $ref: '#/components/examples/JohnRes'
                mary:
                  $ref: '#/components/examples/MaryRes'




  /user/{userId}:
    parameters:
      - in: path
        name: userId
        description: unique id of desired user
        schema:
          type: integer
          format: int64
          minimum: 0
        required: true
    get:
      summary: Gets a user by ID

      responses:
        '200':
          description: Gets the user with the corresponding id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
              examples:
                john:
                  $ref: '#/components/examples/JohnRes'
                mary:
                  $ref: '#/components/examples/MaryRes'

        '404':
          description: User with requested id not found

    put:
      summary: Updates information about user
      requestBody:
        $ref: '#/components/requestBodies/userBody'

      responses:
        '200':
          description: Returns a copy of updated user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
              examples:
                john:
                  $ref: '#/components/examples/JohnRes'
                mary:
                  $ref: '#/components/examples/MaryRes'
        '404':
          description: User with such id does not exist
    delete:
      summary: Deletes the user with desired id if it exists
      responses:
        '204':
          description: User was found and deleted
        '404':
          description: User was not found
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SkyWaet/DataManager/1.0.0